% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ascvd_pooled_coef}
\alias{ascvd_pooled_coef}
\title{ACC/AHA ASCVD Pooled Cohort Equations Coefficients}
\format{
A data structure containing coefficients for ASCVD calculation
}
\source{
CVrisk package
}
\usage{
ascvd_pooled_coef
}
\description{
Coefficient data for calculating ACC/AHA ASCVD risk from the CVrisk package.
These coefficients are used internally by CVrisk functions for risk calculation.
}
\keyword{datasets}
