% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_range}
\alias{check_range}
\title{Check if a value is within a specified range}
\usage{
check_range(x, min, max, var_name)
}
\arguments{
\item{x}{Numeric vector to check}

\item{min}{Minimum value (inclusive)}

\item{max}{Maximum value (inclusive)}

\item{var_name}{Name of the variable for error messages}
}
\value{
Logical vector indicating which values are within range
}
\description{
Check if a value is within a specified range
}
\keyword{internal}
