% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_cholesterol_units}
\alias{convert_cholesterol_units}
\title{Convert cholesterol units from mmol/L to mg/dL}
\usage{
convert_cholesterol_units(value)
}
\arguments{
\item{value}{Numeric vector of cholesterol values in mmol/L}
}
\value{
Numeric vector of cholesterol values in mg/dL
}
\description{
Convert cholesterol units from mmol/L to mg/dL
}
\examples{
convert_cholesterol_units(5.2)  # Returns 201.08
convert_cholesterol_units(1.3)  # Returns 50.27
}
