% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frs_coef}
\alias{frs_coef}
\title{Framingham Risk Score Coefficients}
\format{
A data structure containing coefficients for FRS calculation
}
\source{
CVrisk package
}
\usage{
frs_coef
}
\description{
Coefficient data for calculating Framingham Risk Score from the CVrisk package.
These coefficients are used internally by CVrisk functions for risk calculation.
}
\keyword{datasets}
