% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_scores.R
\name{get_score_friendly_name}
\alias{get_score_friendly_name}
\title{Get friendly name for risk score}
\usage{
get_score_friendly_name(score_col)
}
\arguments{
\item{score_col}{Score column name}
}
\value{
Friendly name for display
}
\description{
Get friendly name for risk score
}
\keyword{internal}
