% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_analysis.R
\name{\%+\%}
\alias{\%+\%}
\title{String concatenation operator}
\usage{
x \%+\% y
}
\arguments{
\item{x}{First string}

\item{y}{Second string}
}
\value{
Concatenated string
}
\description{
String concatenation operator
}
\keyword{internal}
