% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mesa_coef}
\alias{mesa_coef}
\title{MESA CHD Risk Score Coefficients}
\format{
A data structure containing coefficients for MESA calculation
}
\source{
CVrisk package
}
\usage{
mesa_coef
}
\description{
Coefficient data for calculating MESA CHD risk from the CVrisk package.
These coefficients are used internally by CVrisk functions for risk calculation.
}
\keyword{datasets}
