% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_models.R
\name{print.cacs_model_diagnostics}
\alias{print.cacs_model_diagnostics}
\title{Print method for model diagnostics}
\usage{
\method{print}{cacs_model_diagnostics}(x, ...)
}
\arguments{
\item{x}{cacs_model_diagnostics object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for model diagnostics
}
