% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_scores.R
\name{print.ensemble_summary}
\alias{print.ensemble_summary}
\title{Print method for ensemble summary}
\usage{
\method{print}{ensemble_summary}(x, ...)
}
\arguments{
\item{x}{ensemble_summary object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for ensemble summary
}
