% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize}
\alias{standardize}
\title{Standardize variable to mean 0 and SD 1}
\usage{
standardize(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector to standardize}

\item{na.rm}{Logical, whether to remove NA values}
}
\value{
Standardized numeric vector
}
\description{
Standardize variable to mean 0 and SD 1
}
\keyword{internal}
