% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_binary}
\alias{validate_binary}
\title{Validate binary variable}
\usage{
validate_binary(x, var_name)
}
\arguments{
\item{x}{Vector to check}

\item{var_name}{Name of variable for error messages}
}
\value{
Invisible TRUE if valid, otherwise stops with error
}
\description{
Validate binary variable
}
\keyword{internal}
