% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate required columns exist in data}
\usage{
validate_columns(data, required_cols)
}
\arguments{
\item{data}{Data frame to check}

\item{required_cols}{Character vector of required column names}
}
\value{
Invisible TRUE if all columns exist, otherwise stops with error
}
\description{
Validate required columns exist in data
}
\keyword{internal}
