% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_gender}
\alias{validate_gender}
\title{Validate gender variable}
\usage{
validate_gender(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Character vector with standardized gender values
}
\description{
Validate gender variable
}
\keyword{internal}
